Split PreWarm namespace into its own header
This separates the part of PreWarm that inknet needs from the PreWarm
state machine used by http.
