+++ iocore/net/PreWarmManager.h
@@ -23,23 +23,20 @@
 
 #pragma once
 
+#include "PreWarm.h"
 #include "PreWarmAlgorithm.h"
 
 #include "I_EventSystem.h"
 #include "I_NetVConnection.h"
 #include "I_HostDB.h"
-#include "YamlSNIConfig.h"
 #include "NetTimeout.h"
 #include "Milestones.h"
 
 #include "api/Metrics.h"
-#include "records/I_RecHttp.h"
 
 #include <map>
 #include <unordered_map>
-#include <memory>
 #include <queue>
-#include <string_view>
 
 class PreWarmSM;
 class PreWarmManager;
@@ -47,100 +44,6 @@ class SNIConfigParams;
 
 extern ClassAllocator<PreWarmSM> preWarmSMAllocator;
 extern PreWarmManager prewarmManager;
-
-namespace PreWarm
-{
-////
-// Dst
-//
-struct Dst {
-  Dst(std::string_view h, in_port_t p, SNIRoutingType t, int a) : host(h), port(p), type(t), alpn_index(a) {}
-
-  std::string host;
-  in_port_t port      = 0;
-  SNIRoutingType type = SNIRoutingType::NONE;
-  int alpn_index      = SessionProtocolNameRegistry::INVALID;
-};
-
-using SPtrConstDst = std::shared_ptr<const Dst>;
-
-struct DstHash {
-  size_t
-  operator()(const PreWarm::SPtrConstDst &dst) const
-  {
-    CryptoHash hash;
-    CryptoContext context{};
-
-    context.update(dst->host.data(), dst->host.size());
-    context.update(&dst->port, sizeof(in_port_t));
-    context.update(&dst->type, sizeof(SNIRoutingType));
-    context.update(&dst->alpn_index, sizeof(int));
-
-    context.finalize(hash);
-
-    return static_cast<size_t>(hash.fold());
-  }
-};
-
-struct DstKeyEqual {
-  bool
-  operator()(const PreWarm::SPtrConstDst &x, const PreWarm::SPtrConstDst &y) const
-  {
-    return x->host == y->host && x->port == y->port && x->type == y->type && x->alpn_index == y->alpn_index;
-  }
-};
-
-////
-// Conf
-//
-struct Conf {
-  Conf(uint32_t min, int32_t max, double rate, ink_hrtime connect_timeout, ink_hrtime inactive_timeout, bool srv_enabled,
-       YamlSNIConfig::Policy verify_server_policy, YamlSNIConfig::Property verify_server_properties, const std::string &sni)
-    : min(min),
-      max(max),
-      rate(rate),
-      connect_timeout(connect_timeout),
-      inactive_timeout(inactive_timeout),
-      srv_enabled(srv_enabled),
-      verify_server_policy(verify_server_policy),
-      verify_server_properties(verify_server_properties),
-      sni(sni)
-  {
-  }
-
-  uint32_t min                                     = 0;
-  int32_t max                                      = 0;
-  double rate                                      = 1.0;
-  ink_hrtime connect_timeout                       = 0;
-  ink_hrtime inactive_timeout                      = 0;
-  bool srv_enabled                                 = false;
-  YamlSNIConfig::Policy verify_server_policy       = YamlSNIConfig::Policy::UNSET;
-  YamlSNIConfig::Property verify_server_properties = YamlSNIConfig::Property::UNSET;
-  std::string sni;
-};
-
-using SPtrConstConf = std::shared_ptr<const Conf>;
-using ParsedSNIConf = std::unordered_map<SPtrConstDst, SPtrConstConf, DstHash, DstKeyEqual>;
-
-////
-// Stats
-//
-enum class Stat {
-  INIT_LIST_SIZE = 0,
-  OPEN_LIST_SIZE,
-  HIT,
-  MISS,
-  HANDSHAKE_TIME,
-  HANDSHAKE_COUNT,
-  RETRY,
-  LAST_ENTRY,
-};
-
-using StatsIds          = std::array<ts::Metrics::IntType *, static_cast<size_t>(PreWarm::Stat::LAST_ENTRY)>;
-using SPtrConstStatsIds = std::shared_ptr<const StatsIds>;
-using StatsIdMap        = std::unordered_map<SPtrConstDst, SPtrConstStatsIds, DstHash, DstKeyEqual>;
-
-} // namespace PreWarm
 
 /**
    @class PreWarmSM
