+++ iocore/net/TLSTunnelSupport.h
@@ -24,9 +24,10 @@
 
 #pragma once
 
+#include "PreWarm.h"
+
 #include <openssl/ssl.h>
 
-#include "PreWarmManager.h"
 #include "tscore/ink_memory.h"
 #include "tscore/ink_inet.h"
 #include "YamlSNIConfig.h"
